refactor: 로직 hook과 util 파일로 분리
