+++ src/Calculator/Calculator.util.ts
@@ -1,0 +1,24 @@
+import { OPERATION } from './OperationButtonList/OperationButton/OperationButton.type'
+
+export const calculate = ({
+  leftOperand,
+  operation,
+  rightOperand,
+}: {
+  leftOperand: number
+  operation: OPERATION | ''
+  rightOperand: number
+}) => {
+  switch (operation) {
+    case '+':
+      return leftOperand + rightOperand
+    case '-':
+      return leftOperand - rightOperand
+    case '/':
+      return leftOperand / rightOperand
+    case 'X':
+      return leftOperand * rightOperand
+    default:
+      return 0
+  }
+}
